+++ packages/furo-data-input/demo/demos.js
@@ -2,11 +2,13 @@ import "./demo-furo-data-bool-icon";
 import "./sample-furo-data-text-input";
 import "./sample-furo-data-password-input";
 import "./demo-furo-data-text-input";
+import "./demo-furo-data-password-input";
 import "./demo-furo-data-date-input";
+import "./demo-furo-data-search-input";
 import "./demo-furo-data-textarea-input";
 import "./sample-furo-data-number-input"
 import "./demo-furo-data-number-input"
 import "./demo-furo-data-time-input"
-import "./demo-furo-data-time-input"
+import "./demo-furo-data-checkbox-input"
 import "./demo-furo-data-range-input"
 import "./sample-furo-data-collection-dropdown"