+++ packages/furo-data-input/furo-data-date-input.js
@@ -2,6 +2,8 @@ import {LitElement, html, css} from 'lit-element';
 import {Theme} from "@furo/framework/theme"
 import {FBP} from "@furo/fbp";
 import "@furo/input/furo-date-input";
+import {CheckMetaAndOverrides} from "./lib/CheckMetaAndOverrides";
+import {Helper} from "./lib/helper";
 
 /**
  * `furo-data-date-input`
@@ -29,14 +31,105 @@ class FuroDataDateInput extends FBP(LitElement) {
     super();
     this.error = false;
     this.disabled = false;
-    this.errortext = "";
-    this.hint = "";
 
     this._FBPAddWireHook("--valueChanged", (val) => {
+
+      // by valid input reset meta and constraints
+      CheckMetaAndOverrides.CheckAttributeOverrides(this);
+
       if (this.field) {
+
+        if(this.field._spec.type == "google.type.Date") {
+
+          val = this._convertStringToDateObj(val,  this.field.value  );
+        }
+
         this.field.value = val;
       }
     });
+
+    this._FBPAddWireHook("--inputInvalid", (val) => {
+      // val is a ValidityState
+      // https://developer.mozilla.org/en-US/docs/Web/API/ValidityState
+      if (val) {
+        if(val.rangeUnderflow) {
+          this._hint = this._minErrorMessage;
+        }
+        else if(val.rangeOverflow)
+        {
+          this._hint = this._maxErrorMessage;
+        }
+        else if(val.stepMismatch) {
+          this._hint = this._stepErrorMessage;
+        }
+
+        this.requestUpdate();
+      }
+    });
+  }
+
+  /**
+   * Updater for the min => minlength attr*
+   * @param value
+   */
+  set _min(value) {
+    Helper.UpdateInputAttribute(this, "min", value);
+  }
+
+  /**
+   * Updater for the max attr*
+   * @param value
+   */
+  set _max(value) {
+    Helper.UpdateInputAttribute(this, "max", value);
+  }
+
+  /**
+   * Updater for the label attr
+   * @param value
+   */
+  set _label(value) {
+    Helper.UpdateInputAttribute(this, "label", value);
+  }
+
+  /**
+   * Updater for the hint attr
+   * @param value
+   */
+  set _hint(value) {
+    Helper.UpdateInputAttribute(this, "hint", value);
+  }
+
+  /**
+   * Updater for the leadingIcon attr
+   * @param value
+   */
+  set leadingIcon(value) {
+    Helper.UpdateInputAttribute(this, "leading-icon", value);
+  }
+
+  /**
+   * Updater for the trailingIcon attr
+   * @param value
+   */
+  set trailingIcon(value) {
+    Helper.UpdateInputAttribute(this, "trailing-icon", value);
+  }
+
+  /**
+   * Updater for the errortext attr
+   * @param value
+   */
+  set errortext(value) {
+    Helper.UpdateInputAttribute(this, "errortext", value);
+  }
+
+  /**
+   * Updater for the step attr
+   * @param value
+   */
+  set _step(value) {
+    Helper.UpdateInputAttribute(this, "step", value);
   }
 
   static get properties() {
@@ -50,6 +143,14 @@ class FuroDataDateInput extends FBP(LitElement) {
       label: {
         type: String,
         attribute: true
+      },
+      /**
+       * Overrides the required value from the **specs**.
+       *
+       * Use with caution, normally the specs defines this value.
+       */
+      required: {
+        type: Boolean
       },
       /**
        * Overrides the hint text from the **specs**.
@@ -159,83 +260,70 @@ class FuroDataDateInput extends FBP(LitElement) {
    * @param {Object|FieldNode} fieldNode a Field object
    */
   bindData(fieldNode) {
-    if (fieldNode === undefined) {
-      console.warn("Invalid binding ");
-      console.log(this);
-      return
-    }
-
-    this.field = fieldNode;
-    this._updateField();
-    this.field.addEventListener('field-value-changed', (e) => {
-      this._updateField();
-    });
-
-    this.field.addEventListener('field-became-invalid', (e) => {
-      // updates wieder einspielen
-      this.error = true;
-      this.errortext = this.field._validity.description;
-      this.requestUpdate();
-    });
-
-    this.field.addEventListener('field-became-valid', (e) => {
-      // updates wieder einspielen
-      this.error = false;
-      this.requestUpdate();
-    });
+    Helper.BindData(this, fieldNode);
   }
 
 
   _updateField() {
-    // label auf attr ist höher gewichtet
-    if (!this.label) {
-      this._label = this.field._meta.label;
-    } else {
-      this._label = this.label;
-    }
-
-    // hint auf attr ist höher gewichtet
-    if (!this.hint) {
-      this._hint = this.field._meta.hint;
-    } else {
-      this._hint = this.hint;
-    }
-    this.disabled = this.field._meta.readonly ? true : false;
-
-    // min auf attr ist höher gewichtet
-    if (!this.min) {
-      this._min = this.field._meta.min;
-    } else {
-      this._min = this.min;
-    }
-    // max auf attr ist höher gewichtet
-    if (!this.max) {
-      this._max = this.field._meta.max;
-    } else {
-      this._max = this.max;
-    }
-    // step auf attr ist höher gewichtet
-    if (!this.step) {
-      this._step = this.field._meta.step;
-    } else {
-      this._step = this.step;
-    }
-    // readonly auf attr ist höher gewichtet
-    if (!this.readonly) {
-      this._readonly = this.field._meta.readonly;
-    } else {
-      this._readonly = this.readonly;
-    }
-
-
 
     //mark incomming error
     if (!this.field._isValid) {
       this.error = true;
       this.errortext = this.field._validity.description;
     }
-    this._FBPTriggerWire('--value', this.field.value);
+
+    let dateValue = this.field.value;
+
+    // convert value when date type is google.type.Date
+    if(this.field._spec.type == "google.type.Date") {
+      dateValue = this._convertDateObjToString(dateValue);
+    }
+
+    this._FBPTriggerWire('--value', dateValue);
     this.requestUpdate();
+  }
+
+  // convert google date object to ISO 8601
+  _convertDateObjToString(obj) {
+
+    let date ="";
+
+    if(  obj && obj.day && obj.month && obj.year)  {
+      let month = String(obj.month);
+      let day = String(obj.day);
+      let year = String(obj.year);
+
+      if( month.length < 2 ) {
+        month = "0" + month;
+      }
+
+      if( day.length < 2  ) {
+        day = "0" + day;
+      }
+
+      if( year.length < 4  ) {
+        var l = 4-year.length;
+        for(var i=0;  i<l ; i++) {
+          year = "0" + year;
+        }
+      }
+      date = year + "-" + month + "-" + day;
+    }
+
+    return date;
+  }
+
+  // convert date string ISO 8601 to object for google.type.Dates
+  _convertStringToDateObj(str, obj){
+
+    let arr = str.split("-",3);
+    // only override properties: day, month, year
+    if (arr.length === 3) {
+      obj.day = Number(arr[2]);
+      obj.month = Number(arr[1]);
+      obj.year = Number(arr[0]);
+    }
+    return obj;
   }
 
   /**
@@ -264,21 +352,15 @@ class FuroDataDateInput extends FBP(LitElement) {
   render() {
     // language=HTML
     return html` 
-       <furo-date-input 
+       <furo-date-input id="input"  
           ?autofocus=${this.autofocus} 
           ?readonly=${this._readonly||this.disabled} 
-          label="${this._label}" 
-          min="${this._min}" 
-          max="${this._max}" 
-          step="${this._step}" 
           ?error="${this.error}" 
           ?float="${this.float}" 
-          ?condensed="${this.condensed}"          
-          leading-icon="${this.leadingIcon}" 
-          trailing-icon="${this.trailingIcon}" 
-          errortext="${this.errortext}" 
-          hint="${this._hint}" 
+          ?condensed="${this.condensed}"     
+          ?required=${this._required}     
           @-value-changed="--valueChanged"
+          @-input-invalid="--inputInvalid"
           ƒ-set-value="--value"></furo-date-input>      
     `;
   }