+++ packages/furo-data-input/furo-data-range-input.js
@@ -65,9 +65,6 @@ class FuroDataRangeInput extends FBP(LitElement) {
     });
   }
 
-
-
-
   /**
    * Updater for the min => minlength attr*
    * @param value
@@ -128,7 +125,7 @@ class FuroDataRangeInput extends FBP(LitElement) {
    * Updater for the step attr
    * @param value
    */
-  set step(value) {
+  set _step(value) {
     Helper.UpdateInputAttribute(this, "step", value);
   }
 
@@ -252,36 +249,9 @@ class FuroDataRangeInput extends FBP(LitElement) {
    * @param {Object|FieldNode} fieldNode a Field object
    */
   bindData(fieldNode) {
-    if (fieldNode === undefined) {
-      console.warn("Invalid binding ");
-      console.log(this);
-      return
-    }
-    this.field = fieldNode;
-    this._updateField();
-    CheckMetaAndOverrides.UpdateMetaAndConstraints(this);
-    this.field.addEventListener('field-value-changed', (e) => {
-      this._updateField();
-    });
-
-    // update meta and constraints when they change
-    this.field.addEventListener('this-metas-changed', (e) => {
-      CheckMetaAndOverrides.UpdateMetaAndConstraints(this);
-    });
-
-    this.field.addEventListener('field-became-invalid', (e) => {
-      // updates wieder einspielen
-      this.error = true;
-      this.errortext = this.field._validity.description;
-      this.requestUpdate();
-    });
-
-    this.field.addEventListener('field-became-valid', (e) => {
-      // updates wieder einspielen
-      this.error = false;
-      this.requestUpdate();
-    });
+    Helper.BindData(this, fieldNode);
   }
+
 
 
   _updateField() {
@@ -326,9 +296,8 @@ class FuroDataRangeInput extends FBP(LitElement) {
           ?error="${this.error}" 
           ?float="${this.float}" 
           ?condensed="${this.condensed}"          
-          leading-icon="${this.leadingIcon}" 
-          trailing-icon="${this.trailingIcon}" 
           @-value-changed="--valueChanged"
+          @-input-invalid="--inputInvalid"
           ƒ-set-value="--value"></furo-range-input>      
     `;
   }