+++ packages/furo-doc-helper/graph/furo-show-flow.js
@@ -6,20 +6,28 @@ import "./furo-graph-renderer"
 
 /**
  * `furo-show-flow`
- * todo Describe your element
+ * Renders a flow from dom node or html source
  *
- * @summary todo shortdescription
  * @customElement
  * @demo demo/furo-show-flow.html
  * @appliesMixin FBP
  */
 class FuroShowFlow extends FBP(LitElement) {
 
-  constructor() {
-    super();
-
+  /**
+   * Parse html content
+   * @param {string} source
+   */
+  parseHtml(source) {
+    let tpl = document.createElement("div");
+    tpl.innerHTML = source;
+    this.parseTemplate(tpl);
   }
 
+  /**
+   * Parse a dom node
+   * @param {dom} dom node
+   */
   parseTemplate(template) {
     this.graph = new dagre.graphlib.Graph({multigraph: false, compound: true});
     // graph settings